+++ TechTalk.SpecFlow/CucumberMessages/TestResultPartsFactory.cs
@@ -1,29 +1,18 @@
-﻿using System;
-using Io.Cucumber.Messages;
+﻿using Io.Cucumber.Messages;
 using TechTalk.SpecFlow.CommonModels;
-
 using static Io.Cucumber.Messages.TestResult.Types;
 
 namespace TechTalk.SpecFlow.CucumberMessages
 {
-    public interface ITestResultPartsFactory
-    {
-        IResult<TestResult> BuildPassedResult(ulong durationInNanoseconds);
-        IResult<TestResult> BuildFailedResult(ulong durationInNanoseconds, ScenarioContext scenarioContext);
-        IResult<TestResult> BuildAmbiguousResult(ulong durationInNanoseconds, ScenarioContext scenarioContext);
-        IResult<TestResult> BuildPendingResult(ulong durationInNanoseconds, ScenarioContext scenarioContext);
-        IResult<TestResult> BuildSkippedResult(ulong durationInNanoseconds);
-        IResult<TestResult> BuildUndefinedResult(ulong durationInNanoseconds, ScenarioContext scenarioContext, FeatureContext featureContext);
-    }
-
     public class TestResultPartsFactory : ITestResultPartsFactory
     {
+        private readonly ITestAmbiguousMessageFactory _testAmbiguousMessageFactory;
         private readonly ITestErrorMessageFactory _testErrorMessageFactory;
         private readonly ITestPendingMessageFactory _testPendingMessageFactory;
-        private readonly ITestAmbiguousMessageFactory _testAmbiguousMessageFactory;
         private readonly ITestUndefinedMessageFactory _testUndefinedMessageFactory;
 
-        public TestResultPartsFactory(ITestErrorMessageFactory testErrorMessageFactory, ITestPendingMessageFactory testPendingMessageFactory, ITestAmbiguousMessageFactory testAmbiguousMessageFactory, ITestUndefinedMessageFactory testUndefinedMessageFactory)
+        public TestResultPartsFactory(ITestErrorMessageFactory testErrorMessageFactory, ITestPendingMessageFactory testPendingMessageFactory, ITestAmbiguousMessageFactory testAmbiguousMessageFactory,
+            ITestUndefinedMessageFactory testUndefinedMessageFactory)
         {
             _testErrorMessageFactory = testErrorMessageFactory;
             _testPendingMessageFactory = testPendingMessageFactory;
@@ -73,47 +62,5 @@ namespace TechTalk.SpecFlow.CucumberMessages
 
             return Result<TestResult>.Success(testResult);
         }
-
-       
     }
-
-
-    public class TestResultFactory : ITestResultFactory
-    {
-        private readonly ITestResultPartsFactory _testResultPartsFactory;
-
-        public TestResultFactory(ITestResultPartsFactory testResultPartsFactory)
-        {
-            _testResultPartsFactory = testResultPartsFactory;
-        }
-
-        internal ulong ConvertTicksToPositiveNanoseconds(long ticks)
-        {
-            ulong ticksOrZero = (ulong)Math.Min(ticks, 0);
-            return ticksOrZero * 100;
-        }
-
-        public IResult<TestResult> BuildFromContext(ScenarioContext scenarioContext, FeatureContext featureContext)
-        {
-            if (scenarioContext is null)
-            {
-                return Result<TestResult>.Failure(new ArgumentNullException(nameof(scenarioContext)));
-            }
-
-            ulong nanoseconds = ConvertTicksToPositiveNanoseconds(scenarioContext.Stopwatch.Elapsed.Ticks);
-            switch (scenarioContext.ScenarioExecutionStatus)
-            {
-                case ScenarioExecutionStatus.OK: return _testResultPartsFactory.BuildPassedResult(nanoseconds);
-                case ScenarioExecutionStatus.TestError: return _testResultPartsFactory.BuildFailedResult(nanoseconds, scenarioContext);
-                case ScenarioExecutionStatus.StepDefinitionPending: return _testResultPartsFactory.BuildPendingResult(nanoseconds, scenarioContext);
-                case ScenarioExecutionStatus.BindingError: return _testResultPartsFactory.BuildAmbiguousResult(nanoseconds, scenarioContext);
-                case ScenarioExecutionStatus.UndefinedStep: return _testResultPartsFactory.BuildUndefinedResult(nanoseconds, scenarioContext, featureContext);
-                case ScenarioExecutionStatus.Skipped: return _testResultPartsFactory.BuildSkippedResult(nanoseconds);
-                default: return Result<TestResult>.Failure($"Status '{scenarioContext.ScenarioExecutionStatus}' is unknown or not supported.");
-            }
-        }
-    }
-
-    
 }
-
